Add rofi configuration files
This is the first iteration of the files, I might update these later
again as this is a first attempt. The only modes in use currently is the
drun mode for launching applications, and window mode for finding and
switching to open windows. I considered other similar applications such
as wofi, but I like that rofi comes with a window switcher out of the box.
I still might change my mind in the future though. The rofi version I am
configuring is the rofi-wayland fork.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>