Add configurations for wlogout
This just includes 4 options at present. Later I may add more options
e.g. for a soft reboot or a kexec. I may also need to tweak some of the
margins for other devices. But this is a baseline. A somewhat annoying
detail about wlogout is that the default behaviour assumes there will be
3 buttons per row. And will populate placeholders. We work around this
by calling wlogout from the ``power-menu.sh`` script.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>