Add an initial waybar configuration
I've been tinkering with my waybar configuration for a bit and I finally
think I am reasonably happy with what I got now. A floaty status bar
with bright colours for points of interest. It may not be objectively
pretty, but works for now. Even though it might be a bit cluttered.

This commit configures waybar by adding configuration files and scripts
relating to waybar. The modules are organised in
``.config/waybar/modules/``, and the scripts associated with waybar are
placed in ``.config/waybar/scripts/``. The scripts which are waybar
specific are written in python, because I find it easier to work with json
in python compared to bash.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>