Swap the window module for the submap module
Because I find it more useful to have a visual cue to tell me which
submap/mode I am in rather than window. In my opinion the active window
is more apparent. Whilst I am here, I also rearranged the power hover to
be at the right place. And added boxshadows to the taskbar.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>