Add a basic bash script to help me sync changes before commiting
It takes a bit of tweaking until I am happy with the end result of my
configurations. This patch just adds a utility which will sync the files
in the repo with the local files, which will make it simpler to capture
changes. For now, just do this for the home directory.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>