Add a simple install script
This is a super simple install script which will install dependencies
and copies the configurations into the right place. The installer will
copy all files, but in the future it makes sense to update this script
such that it only installs what the user wants. This would be by
prompting the user about what to install. This is good enough for now
though. There may be more dependencies which I've not yet included in
this patch, so the script may be revised in the future.

Signed-off-by: Jekberg <john.berg.96bits@gmail.com>