Add a script to check for missing packages
This is useful as it spits out a list of missing packages if any. And
prints nothing otherwise. All information is printed to stderr so it
should be simple to consume the output of the script in e.g. a pipe. We
also update the install script to use the checkdeps script instead of
doing so itself.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>