Check for yay right before the yay invocation
This just seems to be a more logical place for the yay check. Just
before we execute ``yay -S``. If all packages are installed then we
don't need to exit. Whilst we're here, also update the message if yay is
missing to give some direction on how to install yay.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>