Sync files and directories from /etc
This just makes it easier to get changes from a system into the
repository if they are from the etc directory. Such as system wide
configutation files. There are already a couple of files in the etc
directory but I plan on adding more to allow me to easily reproduce my
environment.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>