Add my current configuration of packages in Portage
This is more or less everything which I really need to setup a new
environment. Some things I don't like e.g. dolphin requires a huge
number of dependencies to work and bloats the system. And using
google-chome and firefox is somewhat redundant. But I will update these
files as I go.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>