EWW a media-player!
Setting up a basic media player with EWW. It is not perfect, but it
allowed me to explore EWW some. There seem to be some odd behaviour in
playerctl e.g. changing position in a track does not update the position
tracking correctly. But I will look into this further later. Now most of
the functionality I want is working. Though I initially attempted to use
the ``image`` widget, it seems much easier to just use a box and set
the background. Because the image seem generates some error logs if a
path is no set, and the playerctl returns a URL rather than a file path.
So it needs to be converted into a file path by stripping the
``file://`` prefix. But a box allows for inline CSS which _can_ handle
URLs.

Some other remarks I have:

- Right now it seems like the player requires redundant ``deflisten`` to
  extract all the necessarity information. There must be a better way to
  get this information.
- I don't really like the background I have currently, but keep it for
  consistency.
- Syling all the components is becoming repetative. Perhaps there is a
  solution to style all components similarly.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>