Build the gentoo kernel with LLVM
The linux kernel supports LTO for building with LLVM. This patch allows
the dist-kernel to be built with the LLVM toolchain. This requires that
the ``LLVM=1`` variable is set and that the appropriate LLVM programs
are set. Otherwise the kernel will fallback to GCC. Some other things
which I noticed were 1) LTO is not working correctly unless LD is
explicitly set to ``ld.lld`, and 2) building the kernel with LTO enabled
break the build unless OBJDUMP, OBJCOPY and READELF are not the LLVM
tooling. I am not uploading a kernel config here though, as this will be
tweaked across devices.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>