Update the hyprpaper config and provide a default wallpaper
On a new install, the background is white from my experiance. This is as
bit painful to look at, so instead make it a pure black background. This
patch contains a minimal ``default.png`` file which is a singel black
pixel. And the new hyprpaper config will use this png file as the
default wallpaper unless another one is set. Also add a nice banner to
the top of the hyprpaper file.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>