Remove the waybar background color and date from the clock
I like a dark background, so the darker background of the waybar is not
needed since all the text is white, and therefore visible agains the
dark background. Also remove the date from the time, as I find this
information to be a bit too much as the current time is mainly what I
need. This also helps as I recently setup a new laptop which has more
limited screen space. So less is more.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>