Add a small scipt and keybind for selecting wallpapers
The ``SUPER + W`` keybind was previously used for hyprexpo, but now I am
repurposing it for wallpaper selection. Image files placed in
``$HOME/.local/wallpepers`` can be selected using the small script added
in this patch. The hyprpaper config will by default set ``default.png``
as the wallpaper which can later be overwritten, but in the future I
would like it to remember the last selected wallpaper on startup. For
now this is a good starting point though.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>