Add a utility script for finding loaded firmware
This is useful when customising Gentoo as it can be used to find the
firmware which has been loaded by the kernel. The script depends on
``journalctl`` but could also work with ``dmesg``. An example of using
the script is listed below:

```
grep -F <(./scripts/list-loaded-firmware.sh) $LINUX_FIRMWARE_CONFIG
```

Where ``$LINUX_FIRMWARE_CONFIG`` is the saved config file. This could be
``/etc/portage/savedconfig/sys-kernel/linux-firmare`` in the case of
Gentoo. The config file can then be updated with the output.

Signed-off-by: Jekberg <john.berg.96bits@gmail.com>