Replace the monitoring modules with the window module
The modules for monitoring cpu, memory and gpu usage currently feel a
bit our of place. Lets replace them with the window module instead as it
makes more sense in the centre. But the current version of the window
module seems to cause other modules to truncate their contents sometimes
which is undesirable, so limit the lenght for now until a better
solution is found.

The gpu-stats.py scritp is removed as part of this commit, but I may
want it again in the future. In that case, I can grab it from the git
history.

Lastly, no need to include radeontop as a dependency. So update the
package sets.

Signed-off-by: Jekberg <john.berg.96bits@gmail.com>