Patch hyprlock to not show a black screen
After upgrading Hyprland to version 0.50, hyprlock has not been taking
the usual screenshots. After building the latest version from source, it
seems that this is fixed in version 0.9.0 or later. Whereas the portage
packae is currently at 0.8.2. Since I'd like to have portage manage the
package updates, cherry-pick the required patches to hyprlock so that
the current version will work. This likely cause build failures when
gui-apps/hyprlock is bumped to 0.9.0 (or later). When this happens, the
patches can be removed.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>