Install the lazy.nvim plugin manager
This is for easy of installing plugins in NeoVim. This seems like a good
starting point as lazy appears to be a popular choice. Maybe I will
change my mind in the future. A low hanging fruit is to get syntax
highlighting going on the gentoo files. This is aleady documented for
lazy, as such lets grab the plugin for Gentoo!

Signed-off-by: Jekberg <john.berg.96bits@gmail.com>