Set up some LSP support for NeoVim
NeoVIM has native supports for LSPs, yet I have yet to tap into this
functionality... unitl now! Turns our that configuring the LSPs are
suprisingly easy as it just requires a config to be created with a given
language server (and some other stuff). I have created a config for
lua-language-server which is an LSP for LUA (who would have thought) as
this is the configuration language foe NeoVIM itself. And I have also
created a config for clangd for C and C++ as I like to write code in
these languages. I provably ought to add some more languages, but this
if sufficient for the time being and there is nothing preventing me from
adding more in the future.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>