Minor LSP tweaks in NeoVIM
Firstly, I have duplicated the enalblement code for the LPS. I added a
simple loop which adds and enables the LSP conigs dynamically so I do not
need it in ``init.lua``. And secondly, the Lua language server fails  to
find the ``vim`` symbol, so it produces some warnings in the file. This
patch fixes both of these small issues

Signed-off-by: John Berg <john.berg.96bits@gmail.com>