Setup TreeSitter and nice highlighting
Until now I have not had any syntax highlighting enabled in NeoVIM. Then
I learned how to enable TreeSitter and switch on configure the
highlighint. What a life chagner.

This patches sets up some highlighting based on the Catppuccin colour
palette. There is plugins available for this but I want to set up the
colour schemes myself because 1) this reduces the need to rely on
plugins, and 2) I can tweak the colour scheme to my liking.

Note that I am still trying to follow the style-guide as-per
https://github.com/catppuccin/catppuccin/blob/main/docs/style-guide.md.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>