Setup tree sitter based folding
I think NeoVIM is using manual folding by default. This is somewhat
annoying to work with as it means using ``zf`` to mstk folds. As a
result I have not used folds much. A much better approach appears to be
to let NeoVIM figure out the folds which lets me then open and close
them as I please. Tree sitter to the rescue! With a few more lines of
Lua, we can setup NeoVIM to fold using the syntax tree it generates.

I am hoping that this will make using folds a more practical solution
going forward. But it would be nice to revisit this later as setting up
folds using LSPs or indentation will also be helpful.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>