Add some files with queries for python and json tree-sitter
There is no syntax highlighting setup for these file types as there is
no query files shipped with NeoVIM by default. Instead of finding an
off-the-shelf solution e.g. a plugin, I think it would be interesting to
learn how this is done. So this patch adds a ``highlights.scm`` file for
python and json. These are likely not complete by any streatch of the
imagination, but they suffice as some of the python files I have at hand
look OK. But I will need to update these files as I go when I see
discrepencies.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>