Split the core set into two components
The core set is now split into ``primary`` which containss  the minimal
number of packages to setup a functioning system (mostly e.g. no kernel
as this is inside the ``kernel`` set). And the ``secondary`` set which
revolves around the graphical packages. This split means that a
functining system can be created even if the system does not have the
requirements.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>