Generate the make.conf file with m4
This is a little bit experimental. The idea is that instead of storing a
dummy ``make.conf`` file in ``etc/make.conf``, we use m4 to generate a
customised configuration file. For now, try to keep this to a minimum as
m4 seems quire powerful. As of this patch ``m4 m4/make.conf.m4`` will
generate a usable portage config wile which is tailored for the host
which it runs on. This leverages the ``recomend-jobs.py`` script to
stamp out lines which can be included in the generated ``make.conf``.

There are some things I may want to change later. For example, require
certain arguments to be provided with the me ``-D`` command line
options. Some aspects also seem a little awkward such as ther being no
convinient way to strip the last new line character of the ``esyscmd()``
function (or this might just be a skill issue) but is seems more effort
than expected.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>