Add an install script for Gentoo
The new ``scripts/gentoo/install.sh`` is an install script for
installing a the minimum number of packages that result in a somewhat
functional system. To keep this simple for now, this excludes anything
gui related and no kernel realted functionality either. The assumptions
is that the script is run after emerge has been used to sync and update
all packages, and it is run on a system created from a stage3 tarball.

I may extend this script later e.g. add command line options to control
the installation to include/exclude stages. But this is not something I
want right now as I have plans on using this minimal install script.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>