Add support for auto completions in NeoVIM
We have LSP support already, but no autocomplete is currently enabled...
Until NOW! All LSPs configured in the ``lsp.lua`` file will be using
autocomplete. The NeoVIM documentation states that using ``noselect`` is
a good option, but I found that this prevents easy completions from
workong e.g. when the autocomplete suggestioin is correct CTRL+Y selects
it. Instead I opted for ``noinsert`` which requies the completion to be
confirmed as described above.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>