Add a new EWW widget which funcions like a control panel
This patch adds a work in progress widget which is accessible from
waybar. It is intended to contain settings and context such as volume
sliders. As I just mentioned, this is not the final iterations. I just
want to test is on multiple systems to get a better feel as to what I
sould put in it. But for now, I also move the media player into this new
widget as it cleans up some of the waybar space.

Also, whilst we are here, lets better organise the files in eww such
that modules go into the modules subdirectory.

Signed-off-by: Jekberg <john.berg.96bits@gmail.com>