Floaty waybar
This changes waybar from being a full bar to be more floaty where
modules are grouped logically and visiualy. But it seems a bit awkward
with to have no gaps when there is only a single window, so add back
some window gaps when we have a single window on a workspace. Also, move
waybar to the bottom layer so that fullscreen still has it visible if
the window which is fullscreened is transpatent (e.g. Alacritty).

Signed-off-by: Jekberg <john.berg.96bits@gmail.com>