Re-organise the the waybar modules
The time now goes to the left and the system tray on the right. This
means there is nothing in the middle for now, but this will soon chagne.
Also merge the logo and the submap as this is not usually useful. And
hide the submap when in the default mode

Signed-off-by: John Berg <john.berg.96bits@gmail.com>