Remove the seconds from the clock module
I have this suspicion that the clock is causing my battery to drain
quicker... as it must update every second. Just rework the clock so we
display the hours and minutes. Also change the tooltip to just be the
date and time (with seconds). Also, also change the text style.

Signed-off-by: Jekberg <john.berg.96bits@gmail.com>