Add a new eww sidebar window
This new window contains buttons for controling the environment e.g.
request a screenshot. And some common shortcuts for e.g. browser. This
sidebar can be toggled by clicking the logo module in waybar.

When I allow someone not familiar with my setup to use it, they seem to
struggle. So I have added a move and kill submap which can be entered
through the new toolbar.

Signed-off-by: Jekberg <john.berg.96bits@gmail.com>