Update the LLVM_SLOT and move it into the llvm use flags fiel
Some packages now support LLVM 21, so time for an upgrade! I also want
to move the LLVM related flags into the llvm package.use flags file.
Also, python is still using an older version of LLVM (and it depends
on the version) so we include python as a special case.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>